app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
qrcode[pil]
python-jose[cryptography]
python-multipart
Pillow
